arquivos/collosfiscal/requirements.txt
streamlit
pandas
lxml
sqlalchemy
psycopg2-binary
python-dotenv